KTIJ-26330 [scripting]: the plugin classloader cannot cope with GradleKotlinScriptConfigurationInputs
The root cause of the issue is that the Kotlin plugin still supports old and new module models.
Components known from xml <include> directives are loaded by the Plugin classloader, whereas those known
from <content> - by module classloaders. When the plugin classloader encounters classes residing in "content"
modules, it signals an error.

When `ScriptConfigurationSnapshotFile` is used to store Gradle file configuration, its
`ScriptConfigurationSnapshotForFS` uses `GradleKotlinScriptConfigurationInputs` (as an implementation of
`CachedConfigurationInputs`). Gradle specifics reside in the `kotlin.gradle.gradle-java` module (part
of the plugin "content"), but components (the class-loading starts from) are available via <include>.
That's the reason of the conflict.

The straightforward solution is to migrate "old" modules to conform to the new model. Unfortunately,
the approach literally assumes the migration of all or nothing. Attempts to stop in the middle failed.
Migration of the plugin modules assumes a lot of class moves and hence external API breakage.

So the current solution is a compromise between huge and painful change and possible (!) performance
degradation in rare cases. `ScriptConfigurationSnapshotFile` is the second cache before s.c.
configuration refinement. It's used only in case the memory cache (the main one) is empty.

^KTIJ-26330 fixed

GitOrigin-RevId: 6554cba468770b6af7f8474dc19a4472c926ff15