[Kotlin] Throttle invalidations of granular caches dropped on low memory
In a low-memory situation, dropping the same cache after receiving multiple
low-memory signals is unlikely to help as the cache in many cases will be
reevaluating the same things over and over again. But it will increase heap
and CPU pressure that might worsen the IDE condition and cause freezes.
Timeout on multiple invalidations of the same cache may help to reduce this
pressure.

KTIJ-27497

GitOrigin-RevId: 10c3bb09e7fec25f7d355e70863e7f0b132b6fe3