[kotlin] Avoid on-air analysis in completion
Together with removing the 'analyzeInDependedAnalysisSession()' call, a few changes in manual PSI handling were required to make completion work again. The reason is that dangling file resolution works a bit differently from on-air analysis.

On-air analysis didn't create a separate 'KtModule' or/and an associated analysis session for in-memory files. Instead, 'LLFirResolveSessionDepended' patched 'getOrBuildFirFor()', so only elements from the re-analyzed FIR subtree were handled specially. In addition, 'resolveToFirSymbol()' worked the similar way – even if one explicitly asked for a 'KtSymbol' for a declaration copy that was not direct subject for re-analysis, a declaration from the original file was returned instead. In some sense, it made the completion implementation a bit chaotic, as one could freely mix 'KtDeclaration's from both the original file and its copy, and still get comparable 'KtSymbol's most of the time.

Dangling file refactoring made the separation between these two files clear. If one requests a 'KtSymbol' from a file copy, they would get precisely it. However, because of the 'IGNORE_SELF' analysis mode that is enabled in completion for performance reasons, 'getContextForPosition()' might contain declarations both from the original file and its copy (if those declarations are local).

So, in certain places where the copy's PSI is manually traversed (for completion item weighting, for instance), 'KtSymbol.equals()' only works correctly if symbols are requested from the same 'KtDeclaration' instance. To simplify mapping between the copy and its origin, the 'getOriginalDeclarationOrSelf()' utility function was added.

GitOrigin-RevId: 70a81edaffe0536bcc862008b2e81c7c02ff0a84