KTIJ-27704 [Scripting]: .gradle.kts files are red until the import is over
The Platform provides an extension point (EP) allowing it to influence the behavior
of highlighting - `ProblemHighlightFilter`. Kotlin plugin uses this EP in particular
to manage highlighting availability (on/off).

Scripts are ready for highlighting once their configuration becomes available.
In the world of Gradle scripts, this means "when the import is over."

GitOrigin-RevId: c083eafa62a0a8cbf1a5bb700014c1dc5d644f7d