[kotlin] Fix library(source) module computation in ambiguous cases
If several libraries share a common JAR, IntelliJ IDEA indexes it only once, so the VirtualFile -> containing library relation is potentially one-to-many.

In cases when a library (also see KT-62933) or a library source file depends on some near file, LLFirFileBuilder is asked to return the dependency. LLFirFileBuilder, which is module-specific, ensures that only files from the same module are queried. It does so by checking if the 'ProjectStructureProvider' returns a consistent result, passing the current module as contextual.

The IDE implementation of ProjectStructureProvider, though, ignores library(source) contextual modules, and relies only on the ProjectFileIndex order (that is arbitrary). As a result, there is a good chance that the first returned library will be the wrong one.

^KTIJ-27977 Fixed

GitOrigin-RevId: 382511fb08959889590608f415b413c6772835fd