IFT-525 Add Java or Kotlin onboarding tour promoter
Created the separate `intellij.kotlin.onboarding-promoter` module for the promoter extension.
This module is not bundled to Kotlin plugin because the promoter should be present only in IDEA Community/Ultimate (when Kotlin and Features Trainer plugins are enabled).
The name of this module is prefixed by `intellij`, because otherwise `KotlinModuleConsistencyTest` will fail. Also, there will be a lot of failed tests because this module won't be loaded in `ClassPathXmlPathResolver.resolveModuleFile` (`intellij` prefixed module descriptors can be missed when loading, by `kotlin` is not).

GitOrigin-RevId: 4b81ff2bfc0b60148e8b91fd35539edd113ab787