[kotlin][debugger] IDEA-338950 produce the exact boxed result for expressions returning an instance of a value class
In such a code fragment there is no need to use the boxed variant of a value class, so the compiler optimizes it, leaving no evidence boxing ever happened in the bytecode. To circumvent the issue, there was an attempt to call toString on an instance. It exposed the information about the value class, but it was just a string, not an actual instance. Wrapping the expression in a cast to Any? does the trick though, it deceives the compiler and the latter, assuming the expression must be boxed here, does not optimize the boxing out.

Other ways to solve the problem are, for example, wrapping it into a generic array and return the first element (might be more tricky for the compiler to optimize this in the future), or just wrap the expression into another, generic value class -- this way, the compiler will optimize out the outer, fake boxing, but leave the inner, untouched value as is. It would require defining such a value class somewhere, probably in stdlib. For now, let's stick to the simplest solution, as long as it works.


Merge-request: IJ-MR-121344
Merged-by: Alexander Kuznetsov <Aleksander.Kuznetsov@jetbrains.com>

GitOrigin-RevId: 3a0841019d84116e2d127ef78298677f921c27ff