[Kotlin] Ignore multiplatform analysis tests with K/N on unsupported hosts
Throw TargetSupportException when kotlin-native-prebuilt can't be
downloaded for the host. Skip tests when the exception is caught.
Because of the JUnit3, tests can't be skipped properly via assumptions.
Due to lack of a better option, the skipped tests will pass as successful
and a warning message will be logged.

KTIJ-28066

GitOrigin-RevId: aa2e9530e8dcae6c5a9b528483e5353468ddaede