[kotlin] change signature: pass declaration as available context for refactoring
In k1, it was an element at caret when available or callable itself. After refactoring for k2: it turned out to be essential to pass an element at caret to correctly resolve applicability. At the same time, default context to detect available types/default values should always be resolved against callable itself.

^KTIJ-28262 fixed

GitOrigin-RevId: 6566e25cf21251fc56a5d4e141735327233cc18a