[debugger] Make "this" parameters visible from the debugger variable view
This commit is a preparation to fix IDEA-340860 problem about exceptions during evaluation of expressions with captured "this" in static lambdas with unnamed parameter. The fix itself is in the next commit.

Before this patch, "this" parameters were not visible from the debugger (as variables). So, the previous approach was to extract values by the StackFrame#getArgumentValues. It leads to exception (see JDK-8323747). As a workaround, JetBrains JDI 2.28 considers "this" parameters as rightful variables in static methods.

Note that code generation for K1 indy mode and for K2 indy mode is different for the singleton object value capturing. In K1 it was a rightful capturing, but K2 does not capture the value and just refers to the global INSTANCE.

GitOrigin-RevId: 0d690203a2cb300d26bec65e67eef1d44627ea32