[kotlin] Reuse applicability checker between extension candidates
Before, auto-completion was based on on-air resolve, and 'FirTowerDataContext' was computed once, during creation of the dependent analysis session (with using 'FirTowerDataContextCollector'). This allowed to return the context almost instantly, no matter how many completion candidates there were.

Dangling file resolution doesn't use 'FirTowerDataContextCollector', as now there is more precise 'ContextCollector'. However, 'ContextCollector' doesn't cache results by itself, and, if asked, it will compute the same context over and over again. This exact happened with extension applicability checker.

The optimized implementation maintains a checker object which caches the computed context. In this commit, the callable contributor is adapted to the change in Analysis API, so now it works more efficiently.

^KTIJ-28445 Fixed

GitOrigin-RevId: 15650f896a08e48bdd776d424d09d3e291ef9ebd