KTIJ-23464 [Scripts]: freezes at configuration reload
Script sdk is a matter of its compilation configuration. The latter is unavailable
until the file is either opened in the editor or gets touched at resolve. Hence,
explicit cache actualization - `getActualClasspathRoots()`, the mean to have desired
script's data in place. Given that configuration refinement is potentially a
time-consuming process, which could be initiated under the platform read lock,
freezes were possible.

This commit stops the practice of the cache actualization and removes the mentioned
`getActualClasspathRoots()` method.

Why taking sdk from "current" cache is safe? We deal with three cases:
1. It's `.gradle.kts`. Configuration is always available after the project import.
2. File is opened in the editor => highlighting. Configuration is always available.
3. File was not opened in the editor => find usages, inspections. Configuration is
   missing. We get default built-ins, and it doesn't seem to affect the functionality.

^KTIJ-23464 fixed

GitOrigin-RevId: 2be2610f9844e9bd9eb07b07229a1717a6d7a5cc