[Kotlin] Added PSI-only based completion that works in dumb mode/during indexing
The PSI-only completion is likely show incorrect or incomplete results, but it should be much better than the text-base completion that was available during indexing so far.
The completion uses a very basic scoping mechanism to understand where declarations are made and available to restrict the shown completions.
It also uses the existing keyword completion to complete Kotlin keywords.

KTIJ-19995

Co-authored-by: Aleksandrina Streltsova <Alexandrina.Streltsova@jetbrains.com>

Merge-request: IJ-MR-126036
Merged-by: Frederik Haselmeier <Frederik.Haselmeier@jetbrains.com>

GitOrigin-RevId: 8f790bad9859fc62c3cc07c7053cf69553990463