[kotlin] Fix `Redundant 'Unit'` inspection after fixing lambda return type inference
KT-62550 highlighted an issue in K2 inference that resulted in the following
incorrect behavior: when a lambda expression that actually returns `Unit` was
passed to a function that expected a non-Unit-returning lambda, the lambda's
return type was inferred to the expected type instead of `Unit`.

This issue has been fixed, and now the type of such a lambda would be correctly
inferred to `(...) -> Unit`. This however, caused the `Redundant 'Unit'` inspection
to give a false positive in the following case:

```kotlin
fun foo(f: () -> Any) {}

fun test() {
    foo {
        return@foo Unit // False positive: Redundant 'Unit'
    }
}
```

(taken from this test:
community/plugins/kotlin/idea/tests/testData/inspectionsLocal/redundantUnitExpression/labeledReturnAny.kt)

Here, since we correctly infer the return type of the lambda to `Unit`, an inspection
is shown.

We fix this by looking at the _expected_ return type of the lambda
(in this case, `Any`), instead of its actual return type (in this case, `Unit`).

This is exactly how it was done in K1
(see `org.jetbrains.kotlin.idea.inspections.RedundantUnitExpressionInspection.Companion#isRedundantUnit`).

GitOrigin-RevId: d807375a70b8fd09a8cccaa1fabd1c8ff61554bb