KTIJ-28552 Can't find built-in class kotlin.Cloneable: ad-hoc solution
`kotlin.Cloneable` absence is caused by missing SDK association for
script's `ModuleInfo`. SDK resolution depends on script metadata, its
definition.

Gradle definitions might be unavailable by the time of a script highlighting.
Current design assumes that some other activity is responsible for their loading
and further synchronization. The reality is that the sync is unreliable.
Our auto tests show, for example, that on Windows hosts without Defender
exclusions for the project, definition loading slows down leading to issues.

This commit introduces an ad-hoc solution. Instead of relying on the cached
definitions, we load them when necessary.

GitOrigin-RevId: f75d8ef9317dfa8bec11d14aaabd94d801f5dfbc