[debugger] KTI-1739 Fix flaky behavior of the resume breakpoint but introduce a problem with step over recursion call
The flaky comes from the fact that the resume breakpoint was set after coroutine went to suspension. But the coroutine can be resumed on another thread before the engine set that resume breakpoint. So to fix it, the resume breakpoint is set now before the Step Over action. But it introduces a problem with recursion calls where the resume breakpoint will be hit.

GitOrigin-RevId: 1fcc4f39450271e76b44cb01b5dceb44e3741d6c