[gradle] IDEA-269370 enable navigation for precompiled script plugins located in `src/main/java`
Gradle documentation mentions that precompiled plugins could be located in `src/main/kotlin` or `src/main/groovy`, depending on the language in which it is written. In practice, Gradle also successfully builds a project where plugins on Kotlin or Groovy are placed inside `src/main/java`. This commit enables navigation for such plugins.

GitOrigin-RevId: 9a0755413778b50a6413d0472f7f8aec0e9de85d