[Kotlin][Java inference] Apply use site scope for j.l.Object
With KMP projects it's possible that a class can reside in a common
module with no JDK available. Trying to use class' resolve scope
for creating java.lang.Object type in this case will result in an
incorrect type. Use site scope from a type comes from a JVM module
which has a JDK available.

KTIJ-28461

GitOrigin-RevId: b19fe424ce4f7745e7f6aad7c5a34bb3912f5a56