[kotlin] AMPER-659 Group opt-in quickfixes together for better readability
The low-prio status of the module-wide opt-in quickfix is not that important
compared to the added clarity of grouping all opt-in quickfixes at different
levels (statement, function, file, AND module).
This is why we decided to bring the module-wide opt-in to normal prio so it
appears next to its friends.

GitOrigin-RevId: f370b869fdd215cfddb47c0ab2d3402b3ca5ab0f