KTIJ-30040 [kotlin] Ignore suspension calls in non-inlined local declarations in `RedundantSuspendModifierInspection`
Only the expressions which are executed directly in the function's body
should be analyzed when deciding whether the `suspend` modifier
is redundant or not.

The local declarations can have their own `suspend` modifiers,
and are independent of the containing declaration, with the
exception for inlined lambdas.

Add more flexible `KotlinCallProcessor.process` overload,
which takes a lazily produced sequence.

Remove `processExpressionsRecursively`, which was too
rigid, and was eagerly collecting all descendants
into a list.

^KTIJ-30040 Fixed

GitOrigin-RevId: b585c864c9c04e924e30e8a09d968500f6129962