[coroutines-debugger]: Fix Step Out of suspend functions
Before stepping, we visit the state machine of the suspend function and find 2 instructions: `areturn` that follows coroutine suspension (`firstReturnAfterSuspensionOffset`) and the first instruction that will be executed after suspension (`nextCallOffset`). Resume breakpoint is set to the first line of invokeSuspend and the regular step out starts.

If `KotlinStepOutRequestHint` reaches `firstReturnAfterSuspensionOffset` with `STEP_MIN` from the current position -> the coroutine was suspended -> resume the execution and wait for the resume breakpoint -> step till `nextCallOffset` -> `STOP`.
Otherwise, if the coroutine was not suspended -> proceed with the regular step out.


Merge-request: IJ-MR-136542
Merged-by: Maria Sokolova <maria.sokolova@jetbrains.com>

GitOrigin-RevId: d4f74d70c63f4ef2a538c63f70f183acad7e2135